requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.8=h166bdaf_0
asttokens=2.0.5=pyhd3eb1b0_0
attr=2.5.1=h166bdaf_1
backcall=0.2.0=pyhd3eb1b0_0
binutils_impl_linux-64=2.39=he00db2b_1
binutils_linux-64=2.39=h5fc0e48_13
boost-cpp=1.78.0=h6582d0a_3
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.5.7=hbcca054_0
cairo=1.16.0=hbbf8b49_1016
certifi=2023.5.7=pyhd8ed1ab_0
charset-normalizer=3.1.0=pyhd8ed1ab_0
cmake=3.26.4=pypi_0
comm=0.1.2=py311h06a4308_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
debugpy=1.6.7=py311hcafe171_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.4.0=h4bd325d_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=hcb278e6_1
fenics=2019.1.0=py311h38be061_34
fenics-dijitso=2019.1.0=py311h38be061_34
fenics-dolfin=2019.1.0=py311h0a5b7eb_34
fenics-ffc=2019.1.0=py311h38be061_34
fenics-fiat=2019.1.0=py311h38be061_34
fenics-libdolfin=2019.1.0=h0f808bc_34
fenics-ufl=2019.1.0=py311h38be061_34
fftw=3.3.10=mpi_mpich_h5537406_7
filelock=3.12.2=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.39.4=pypi_0
freetype=2.12.1=hca18f0e_1
gcc_impl_linux-64=10.4.0=h5231bdf_19
gcc_linux-64=10.4.0=h9215b83_13
gettext=0.21.1=h27087fc_0
glib=2.76.4=hfc55251_0
glib-tools=2.76.4=hfc55251_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.22.3=he621ea3_0
gstreamer=1.22.3=h5eee18b_0
gxx_impl_linux-64=10.4.0=h5231bdf_19
gxx_linux-64=10.4.0=h6e491c6_13
h5py=3.8.0=pypi_0
harfbuzz=7.3.0=hdb3a94d_0
hdf5=1.12.2=mpi_mpich_h5d83325_1
hypre=2.25.0=mpi_mpich_hed3a557_0
icu=72.1=hcb278e6_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
importlib_metadata=6.6.0=hd8ed1ab_0
ipykernel=6.23.1=pyh210e3f2_0
ipython=8.12.0=py311h06a4308_0
jax=0.4.13=pypi_0
jaxlib=0.4.13=pypi_0
jedi=0.18.1=py311h06a4308_1
jinja2=3.1.2=pypi_0
jupyter_client=8.1.0=py311h06a4308_0
jupyter_core=5.3.0=py311h06a4308_0
kernel-headers_linux-64=2.6.32=he073ed8_15
kiwisolver=1.4.4=py311h4dd048b_1
krb5=1.20.1=hf590b9c_0
lame=3.100=h166bdaf_1003
lcms2=2.15=haa2dc70_1
ld_impl_linux-64=2.39=hcc3a1bd_1
lerc=4.0.0=h27087fc_0
libabseil=20230125.2=cxx17_h59595ed_2
libaec=1.0.6=hcb278e6_1
libblas=3.9.0=17_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcap=2.67=he9d0100_0
libcblas=3.9.0=17_linux64_openblas
libclang=16.0.6=default_h1cdf331_0
libclang13=16.0.6=default_h4d60ac6_0
libcups=2.3.3=h36d4200_3
libcurl=7.88.1=h251f7ec_1
libdeflate=1.18=h0b41bf4_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hf998b51_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libflac=1.4.3=h59595ed_0
libgcc-devel_linux-64=10.4.0=hd38fd1e_19
libgcc-ng=13.1.0=he5830b7_0
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.4=hebfc3b9_0
libgomp=13.1.0=he5830b7_0
libgpg-error=1.47=h71f35ed_0
libgrpc=1.54.2=hb20ce57_2
libiconv=1.17=h166bdaf_0
libjpeg-turbo=2.1.5.1=h0b41bf4_0
liblapack=3.9.0=17_linux64_openblas
libllvm16=16.0.6=h5cf9203_0
libnghttp2=1.52.0=h2d74bed_0
libnsl=2.0.0=h5eee18b_0
libogg=1.3.5=h27cfd23_1
libopenblas=0.3.23=pthreads_h80387f5_0
libopus=1.3.1=h7f98852_1
libpng=1.6.39=h753d276_0
libpq=15.3=hbcd7760_1
libprotobuf=3.21.12=h3eb15da_0
libsanitizer=10.4.0=h5246dfb_19
libsndfile=1.2.0=hb75c966_0
libsodium=1.0.18=h7b6447c_0
libsqlite=3.42.0=h2797004_0
libssh2=1.10.0=hdbd6064_1
libstdcxx-devel_linux-64=10.4.0=hd38fd1e_19
libstdcxx-ng=13.1.0=hfd8a6a1_0
libsystemd0=253=h8c4010b_1
libtiff=4.5.1=h8b53f26_0
libuuid=2.38.1=h0b41bf4_0
libuv=1.44.2=h5eee18b_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.1=hd590300_0
libxcb=1.15=h0b41bf4_0
libxkbcommon=1.5.0=h5d7e998_3
libxml2=2.11.4=h0d562d8_0
libzlib=1.2.13=h166bdaf_4
lit=16.0.6=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-base=3.7.2=py311h54ef318_0
matplotlib-inline=0.1.6=py311h06a4308_0
metis=5.1.0=hf484d3e_4
ml_dtypes=0.2.0=py311h320fe9a_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.2.0=hb012696_0
mpg123=1.31.3=hcb278e6_0
mpi=1.0=mpich
mpi4py=3.1.4=py311h7edb0b5_0
mpich=4.0.3=h846660c_100
mpmath=1.2.1=py311h06a4308_0
mumps-include=5.2.1=ha770c72_11
mumps-mpi=5.2.1=h7ee95aa_11
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.33=hf1915f5_2
mysql-libs=8.0.33=hca2cd23_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py311h06a4308_0
networkx=3.1=pypi_0
nspr=4.35=h27087fc_0
nss=3.89=he45b914_0
numpy=1.24.3=py311h64a7726_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.1=hd590300_1
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=23.0=py311h06a4308_0
parmetis=4.0.3=h2a9763c_1005
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.40=hc3806b6_0
petsc=3.17.4=real_h15390b8_101
petsc4py=3.17.4=real_h9c72440_101
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py311h06a4308_0
pixman=0.40.0=h36c2ea0_0
pkg-config=0.29.2=h1bed415_8
pkgconfig=1.5.4=py311h06a4308_0
platformdirs=2.5.2=py311h06a4308_0
ply=3.11=py_1
pooch=1.7.0=pyha770c72_3
prompt-toolkit=3.0.36=py311h06a4308_0
psutil=5.9.0=py311h5eee18b_0
pthread-stubs=0.4=h36c2ea0_1001
ptscotch=6.0.9=hb499603_2
ptyprocess=0.7.0=pyhd3eb1b0_2
pulseaudio-client=16.1=hb77b528_4
pure_eval=0.2.2=pyhd3eb1b0_0
pybind11=2.10.1=py311hdb19cb5_0
pybind11-global=2.10.1=py311hdb19cb5_0
pygments=2.15.1=py311h06a4308_1
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py311ha74522f_3
pyqt5-sip=12.11.0=py311hcafe171_3
pysocks=1.7.1=pyha2e5f31_6
python=3.11.4=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.11=3_cp311
pyzmq=25.1.0=py311h6a678d5_0
qt-main=5.15.8=h01ceb2d_12
re2=2023.03.02=h8c504da_0
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
rhash=1.4.3=hdbd6064_0
scalapack=2.2.0=hd931219_1
scipy=1.10.1=py311h64a7726_3
scotch=6.0.9=hb2e6521_2
setuptools=67.8.0=py311h06a4308_0
sip=6.7.9=py311hb755f60_0
six=1.16.0=pyhd3eb1b0_1
slepc=3.17.2=real_hb19a63e_100
slepc4py=3.17.2=real_h97905d6_101
stack_data=0.2.0=pyhd3eb1b0_0
suitesparse=5.10.1=h9e50725_1
superlu=5.2.2=h00795ac_0
superlu_dist=7.2.0=h25dcc4a_0
sympy=1.11.1=py311h06a4308_0
sysroot_linux-64=2.12=he073ed8_15
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torch=2.0.1=pypi_0
tornado=6.2=py311h5eee18b_0
traitlets=5.7.1=py311h06a4308_0
triton=2.0.0=pypi_0
typing-extensions=4.7.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.3=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py311h06a4308_0
xcb-util=0.4.0=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.39=hd590300_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.6=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xf86vidmodeproto=2.3.1=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.5=hc292b87_0
